# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: API REST do Espaço Terapia OS (clínica pediátrica multi-filial)
  title: Espaço Terapia API
  version: "1.0"
paths:
...
  /notification-settings:
    get:
      parameters:
      - description: Unidade (UUID)
        in: query
        name: unidade_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obter configurações de notificação
      tags:
      - notification
    put:
      consumes:
      - application/json
      parameters:
      - description: Unidade (UUID)
        in: query
        name: unidade_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Salvar configurações de notificação
      tags:
      - notification
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"